requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
pymupdf
requests
huggingface_hub
